+++ python2/raygun4py/raygunmsgs.py
@@ -142,5 +142,16 @@ class RaygunErrorMessage:
 
         if '__traceback_hide__' not in localVars:
             for key in localVars:
-                result[key] = str(localVars[key])
+                try:
+                    # Note that str() *can* fail; thus protect against it as much as we can.
+                    if type(localVars[key]) is unicode:
+                        result[key] = localVars[key]
+                    else:
+                        result[key] = str(localVars[key])
+                except Exception as e:
+                    try:
+                        r = repr(localVars[key])
+                    except Exception as re:
+                        r = "Couldn't convert to repr due to {0}".format(re)
+                    result[key] = "!!! Couldn't convert {0!r} (repr: {1}) due to {2!r} !!!".format(key, r, e)
             return result