+++ python3/raygun4py/raygunmsgs.py
@@ -1,4 +1,3 @@
-import traceback
 import inspect
 import os
 
@@ -155,5 +154,13 @@ class RaygunErrorMessage:
 
         if '__traceback_hide__' not in localVars:
             for key in localVars:
-                result[key] = str(localVars[key])
+                try:
+                    # Note that str() *can* fail; thus protect against it as much as we can.
+                    result[key] = str(localVars[key])
+                except Exception as e:
+                    try:
+                        r = repr(localVars[key])
+                    except Exception as re:
+                        r = "Couldn't convert to repr due to {0}".format(re)
+                    result[key] = "!!! Couldn't convert {0!r} (repr: {1}) due to {2!r} !!!".format(key, r, e)
             return result